add SweetAlert confirm delete

+++ src/components/Product.jsx
@@ -2,6 +2,7 @@ import React from "react";
 import { Link } from "react-router-dom";
 import { useDispatch } from "react-redux";
 import { deleteProductAction } from "../actions/productAction";
+import Swal from "sweetalert2";
 
 export const Product = ({ product }) => {
   const { name, price, id } = product;
@@ -9,7 +10,20 @@ export const Product = ({ product }) => {
   const dispatch = useDispatch();
 
   const confirmDeleteProduct = (id) => {
-    dispatch(deleteProductAction(id));
+    Swal.fire({
+      title: "Estas seguro?",
+      text: "Un producto que se elimina no se puede recuperar!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Si, eliminar!",
+      cancelButtonText: "Cancelar",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        dispatch(deleteProductAction(id));
+      }
+    });
   };
 
   return (
